requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
google-generativeai
jinja2
aiofiles
pytest
httpx 
python-dotenv